Update comment.ts
It seems like the ESLint is still producing a warning for the variable _unusedKey being defined but never used. To address this warning in ESLint, you can disable the @typescript-eslint/no-unused-vars rule for that specific line using an eslint comment:

By commenting out _unusedKey within the array destructuring, you are essentially telling ESLint to ignore this variable for the purpose of the no-unused-vars rule.